new function added to crop data to geography
I frequently have to help BirdLife partners in various countries to identify IBAs in their respective country. Hence I need to crop tracking data to retain only data for a certain country (or EEZ). This function does that and also re-labels the track IDs so that any interpolation of those data does not add lots of points to the boundary of the geography (e.g. if the same bird leaves and then re-enters the geography after 5 months).